etl/requirements.txt
python-dotenv
PyYAML
requests
ijson
sqlalchemy
pandas
psycopg2-binary
jupyter
